requirements.txt
Pillow==9.3.0
decord==0.6.0
gradio==4.31.5
ninja==1.11.1.1
omegaconf==2.3.0
openai==1.14.2
pathos==0.3.2
prettytable==3.10.0
protobuf==3.20.3
pycocoevalcap==1.2
pycocotools==2.0.8
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
tiktoken==0.6.0
torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu121
https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.7/flash_attn-2.5.7+cu122torch2.1cxx11abiFALSE-cp39-cp39-linux_x86_64.whl
transformers==4.37.2
triton==2.2.0
